Added Moving X Axis Camera
The camera is not finished yet as it can go off screen.  Will add
barrier later to where the camera can go.

Also added animation variables to check if the character is grounded for
jumping.  Can't jump if the character isn't on the ground.  Jumping is
NOT implemented yet, just the variable for grounded.