Timer has now been styled, scales
Right now, the timer's font size scales based on the timer box's current
width. I may change how this is calculated at a later time, as I'd like
to factor in both the width and height.

One possible solution would be to just make both the X and Y dimensions
of the box size the exact same size, so this may happen.